Divide files
